requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.6
    # via kombu
aniso8601==7.0.0
    # via graphene
anyascii==0.2.0
    # via wagtail
arrow==1.1.0
    # via -r requirements.in
asgiref==3.3.4
    # via django
attrs==20.3.0
    # via
    #   fiona
    #   pytest
beautifulsoup4==4.8.2
    # via wagtail
billiard==3.6.4.0
    # via celery
celery==5.0.5
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   fiona
    #   pyproj
    #   requests
    #   sentry-sdk
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   fiona
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   fiona
click-repl==0.1.6
    # via celery
cligj==0.7.1
    # via fiona
cycler==0.10.0
    # via matplotlib
django==3.1.9
    # via
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-modeltrans
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   graphene-django
    #   wagtail
    #   wagtail-localize
django-cors-headers==3.7.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==2.4.0
    # via wagtail
django-modelcluster==5.1
    # via wagtail
django-modeltrans==0.6.0
    # via -r requirements.in
django-prometheus==2.1.0
    # via -r requirements.in
django-taggit==1.4.0
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
factory-boy==3.2.0
    # via
    #   -r requirements.in
    #   pytest-factoryboy
    #   wagtail-factories
faker==8.1.2
    # via factory-boy
filterpy==1.4.5
    # via -r requirements.in
fiona==1.8.21
    # via geopandas
fonttools==4.36.0
    # via matplotlib
geopandas==0.11.1
    # via -r requirements.in
graphene==2.1.9
    # via
    #   -r requirements.in
    #   graphene-django
graphene-django==2.15.0
    # via -r requirements.in
graphene-django-optimizer==0.8.0
    # via -r requirements.in
graphql-core==2.3.2
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via graphene
greenlet==1.1.2
    # via sqlalchemy
html5lib==1.1
    # via wagtail
idna==2.10
    # via requests
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
jinja2==3.0.1
    # via -r requirements.in
kiwisolver==1.3.1
    # via matplotlib
kombu==5.0.2
    # via celery
l18n==2020.6.1
    # via wagtail
llvmlite==0.39.0
    # via numba
markupsafe==2.0.1
    # via jinja2
matplotlib==3.5.3
    # via filterpy
munch==2.5.0
    # via fiona
numba==0.56.0
    # via -r requirements.in
numpy==1.22.4
    # via
    #   filterpy
    #   matplotlib
    #   numba
    #   pandas
    #   scipy
openpyxl==3.0.7
    # via tablib
orjson==3.5.2
    # via -r requirements.in
owslib==0.25.0
    # via -r requirements.in
packaging==20.9
    # via
    #   geopandas
    #   matplotlib
    #   pytest
pandas==1.4.3
    # via
    #   -r requirements.in
    #   geopandas
patchy==2.2.0
    # via -r requirements.in
pillow==8.2.0
    # via
    #   matplotlib
    #   wagtail
pluggy==0.13.1
    # via pytest
polib==1.1.1
    # via wagtail-localize
prometheus-client==0.11.0
    # via django-prometheus
promise==2.3
    # via
    #   graphene-django
    #   graphql-core
    #   graphql-relay
prompt-toolkit==3.0.18
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyproj==3.0.1
    # via
    #   geopandas
    #   owslib
pytest==6.2.4
    # via
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-lazy-fixture
pytest-django==4.2.0
    # via -r requirements.in
pytest-factoryboy==2.1.0
    # via -r requirements.in
pytest-lazy-fixture==0.6.3
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   arrow
    #   faker
    #   matplotlib
    #   owslib
    #   pandas
python-dotenv==0.17.1
    # via -r requirements.in
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-modelcluster
    #   l18n
    #   owslib
    #   pandas
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   owslib
ranking==0.3.2
    # via -r requirements.in
redis==3.5.3
    # via -r requirements.in
requests==2.25.1
    # via
    #   owslib
    #   responses
    #   wagtail
responses==0.13.3
    # via -r requirements.in
rx==1.6.1
    # via graphql-core
scipy==1.9.0
    # via
    #   -r requirements.in
    #   filterpy
sentry-sdk==1.0.0
    # via -r requirements.in
shapely==1.7.1
    # via geopandas
singledispatch==3.6.1
    # via graphene-django
six==1.15.0
    # via
    #   click-repl
    #   cycler
    #   fiona
    #   graphene
    #   graphene-django
    #   graphql-core
    #   graphql-relay
    #   html5lib
    #   l18n
    #   munch
    #   promise
    #   python-dateutil
    #   responses
    #   singledispatch
soupsieve==2.2.1
    # via beautifulsoup4
sqlalchemy==1.4.40
    # via -r requirements.in
sqlparse==0.4.1
    # via django
tablib[xls,xlsx]==3.0.0
    # via wagtail
text-unidecode==1.3
    # via
    #   faker
    #   graphene-django
toml==0.10.2
    # via pytest
urllib3==1.26.4
    # via
    #   requests
    #   responses
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
wagtail==2.12.4
    # via
    #   -r requirements.in
    #   wagtail-factories
    #   wagtail-localize
wagtail-factories==2.0.1
    # via -r requirements.in
wagtail-localize==0.9.5
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow==1.4
    # via wagtail
xlrd==2.0.1
    # via tablib
xlsxwriter==1.4.0
    # via wagtail
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
